Fix merge conflict residue
