+++ src/app/game/car/car.scss
@@ -1,0 +1,24 @@
+:host {
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.controls {
+  display: none;
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(255,255,255,0.6);
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
+:host(:hover) .controls {
+  display: flex;
+}
+button {
+  margin: 2px;
+}
